Refactor doBoundsIntersect
