+++ packages/element/src/collision.ts
@@ -11,7 +11,6 @@ import {
   vectorFromPoint,
   vectorNormalize,
   vectorScale,
-  doBoundsIntersect,
 } from "@excalidraw/math";
 
 import {
@@ -26,6 +25,7 @@ import type { FrameNameBounds } from "@excalidraw/excalidraw/types";
 import { isPathALoop } from "./utils";
 import {
   type Bounds,
+  doBoundsIntersect,
   elementCenterPoint,
   getCenterForBounds,
   getElementBounds,
