+++ packages/excalidraw/lasso/utils.ts
@@ -4,12 +4,12 @@ import {
   polygonFromPoints,
   lineSegment,
   polygonIncludesPointNonZero,
-  doBoundsIntersect,
 } from "@excalidraw/math";
 
 import {
   type Bounds,
   computeBoundTextPosition,
+  doBoundsIntersect,
   getBoundTextElement,
   getElementBounds,
   intersectElementWithLineSegment,
